Fixed quiet a few things for #21
https://github.com/Meshbits/pirate-wallet-layouts/issues/21

'Change PIN' must be a part of the Nav-bar. Also, reduce the size as per design
In Top Nav: The back button is missing.
Instead of the 'Change PIN' text, I think only 'PIN' is sufficient on this screen because it's the first time for a user to set up a new PIN.
Avoid using ALL CAPS in the title. Fix 'Set' text
Move Entire middle content (including 'Set PIN' title, to the top.
Fix font color & size, specially in sub-title & supportive text.
keypad - fixed text isn't in center